Matching Problem
In a matching problem, two or more points are matched and total distance of these matches is targetted to minimize. In this example, we have 3 models. First model aims to minimize total distance, the second aims to minimize maximum distance between particular points, and last model aims to firstly minimize maximum distance and then minimize total distance of matches by depending on first objective. The last model is a multi-objective model.