Employee Scheduling
Employee scheduling explained in this notebook is not a hard problem. Of course, you can face a very challenging problem regarding employee scheduling (workforce optimization). In this model, you have to solve a 2-day scheduling problem (suppose it is valentine's day and you made some campaigns for this special day). You used historical data (past valentine's day) and predicted employee requirements for every 2-hour period. Every employee can be worked at most 8 hours because of union or government rules. How can you optimize this process? The trickiest part is identifying the appropriate decision variables. You need to identify the number or workers who start to work as variables in a particular period. If you choose the number of employees working in a particular period as a variable, you will absolutely fail.